requirements.txt
certifi==2018.11.29
chardet==3.0.4
idna==2.8
mysql-connector-python==8.0.15
numpy==1.16.1
oauthlib==3.0.1
pandas==0.24.1
protobuf==3.6.1
python-dateutil==2.8.0
pytz==2018.9
requests==2.21.0
requests-oauthlib==1.2.0
six==1.12.0
twython==3.7.0
Unidecode==1.0.23
urllib3==1.24.1
